opspilot-enterprise/services/langgraph-orchestrator/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "opspilot-langgraph-orchestrator"
version = "0.1.0"
description = "OpsPilot LangGraph Orchestrator mock service"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "httpx",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "opspilot-shared-schema @ file:../../packages/shared-schema",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
